Conflicts:
	shell.py

+++ seqr/management/commands/update_projects_in_new_schema.py
@@ -3,8 +3,9 @@ import json
 import logging
 import pymongo
 from tqdm import tqdm
-
 import settings
+
+
 from django.core.management.base import BaseCommand
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import Q
@@ -31,10 +32,13 @@ from seqr.models import \
     VariantTagType as SeqrVariantTagType, \
     VariantTag as SeqrVariantTag, \
     VariantNote as SeqrVariantNote, \
-    SequencingSample as SeqrSequencingSample, \
+    Sample as SeqrSample, \
     SampleBatch as SeqrSampleBatch, \
     LocusList, \
-    CAN_EDIT, CAN_VIEW
+    CAN_EDIT, CAN_VIEW, ModelWithGUID
+
+from xbrowse_server.mall import get_datastore, get_annotator
+
 
 logger = logging.getLogger(__name__)
 
@@ -66,15 +70,15 @@ class Command(BaseCommand):
 
         if reset_all_models:
             print("Dropping all records from SeqrProject, SeqrFamily, SeqrIndividual")
-            SeqrProject.objects.all().delete()
-            SeqrFamily.objects.all().delete()
             SeqrIndividual.objects.all().delete()
+            SeqrFamily.objects.all().delete()
+            SeqrProject.objects.all().delete()
 
         # reset models that'll be regenerated
         SeqrVariantTagType.objects.all().delete()
         SeqrVariantTag.objects.all().delete()
         SeqrVariantNote.objects.all().delete()
-        SeqrSequencingSample.objects.all().delete()
+        SeqrSample.objects.all().delete()
         SeqrSampleBatch.objects.all().delete()
 
 
@@ -104,17 +108,19 @@ class Command(BaseCommand):
 
             print("Project: " + source_project.project_id)
 
-            # compute sequencing_type for this project
+
+            # compute sample_type for this project
             project_names = ("%s|%s" % (source_project.project_id, source_project.project_name)).lower()
             if "wgs" in project_names or "genome" in source_project.project_id.lower() or source_project.project_id.lower() in wgs_project_ids:
-                sequencing_type = SeqrSampleBatch.SEQUENCING_TYPE_WGS
+                sample_type = SeqrSampleBatch.SAMPLE_TYPE_WGS
                 counters['wgs_projects'] += 1
             elif "rna-seq" in project_names:
-                sequencing_type = SeqrSampleBatch.SEQUENCING_TYPE_RNA
+                sample_type = SeqrSampleBatch.SAMPLE_TYPE_RNA
                 counters['rna_projects'] += 1
             else:
-                sequencing_type = SeqrSampleBatch.SEQUENCING_TYPE_WES
+                sample_type = SeqrSampleBatch.SAMPLE_TYPE_WES
                 counters['wes_projects'] += 1
+
 
             # transfer Project data
             new_project, project_created = transfer_project(source_project)
@@ -144,31 +150,41 @@ class Command(BaseCommand):
                     if individual_created: counters['individuals_created'] += 1
                     if phenotips_data_retrieved: counters['individuals_data_retrieved_from_phenotips'] += 1
 
-                    vcf_files = [f for f in source_individual.vcf_files.all()]
-                    vcf_path = None
-                    if len(vcf_files) > 0:
-                        # get the most recent VCF file (the one with the highest primary key
-                        vcf_files_max_pk = max([f.pk for f in vcf_files])
-                        vcf_path = [f.file_path for f in vcf_files if f.pk == vcf_files_max_pk][0]
 
-                    if vcf_path:
-                        new_sample_batch, sample_batch_created = get_or_create_sample_batch(
-                            new_project,
-                            path=vcf_path,
-                            sequencing_type=sequencing_type,
-                            genome_build_id=GENOME_BUILD_GRCh37,
-                        )
+                    if source_individual.combined_individuals_info:
+                        combined_individuals_info = json.loads(source_individual.combined_individuals_info)
+                        """
+                        combined_individuals_info json is expected to look like:
+                        {
+                            'WES' : {
+                                'project_id': from_project.project_id,
+                                'family_id': from_f.family_id,
+                                'indiv_id': from_i.indiv_id
+                            },
+                            'WGS' : {
+                                'project_id': from_project.project_id,
+                                'family_id': from_f.family_id,
+                                'indiv_id': from_i.indiv_id
+                            },
+                            'RNA' : {
+                                'project_id': from_project.project_id,
+                                'family_id': from_f.family_id,
+                                'indiv_id': from_i.indiv_id
+                            },
+                        }
+                        """
+                        for i, sample_type_i, combined_individuals_info_i in enumerate(combined_individuals_info.items()):
+                            source_project_i = Project.objects.get(project_id=combined_individuals_info_i['project_id'])
+                            #source_family_i = Project.objects.get(project_id=combined_individuals_info_i['family_id'])
+                            source_individual_i = Project.objects.get(project_id=combined_individuals_info_i['indiv_id'])
 
-                        sample, sample_created = get_or_create_sample(
-                            source_individual,
-                            new_sample_batch,
-                            new_individual,
-                        )
-
-                    if sample_created: counters['samples_created'] += 1
+                            create_sample_records(sample_type_i, source_project_i, source_individual_i, new_project, new_individual, counters)
+                    else:
+                        create_sample_records(sample_type, source_project, source_individual, new_project, new_individual, counters)
+                        #combined_families_info.update({from_project_datatype: {'project_id': from_project.project_id, 'family_id': from_f.family_id}})
 
             # TODO family groups, cohorts
-            for source_variant_tag_type in ProjectTag.objects.filter(project=source_project):
+            for source_variant_tag_type in ProjectTag.objects.filter(project=source_project).order_by('order'):
                 new_variant_tag_type, created = get_or_create_variant_tag_type(
                     source_variant_tag_type, new_project)
 
@@ -195,32 +211,61 @@ class Command(BaseCommand):
 
 
         # delete projects that are in SeqrIndividual table, but not in BaseProject table
-        for indiv in SeqrIndividual.objects.all():
-            if indiv.guid not in updated_seqr_individual_guids:
-                print("Deleting SeqrIndividual: %s" % indiv)
-                indiv.delete()
+        if not project_ids_to_process:
+            for indiv in SeqrIndividual.objects.all():
+                if indiv.guid not in updated_seqr_individual_guids:
+                    print("Deleting SeqrIndividual: %s" % indiv)
+                    indiv.delete()
 
-        # delete projects that are in SeqrFamily table, but not in BaseProject table
-        for f in SeqrFamily.objects.all():
-            if f.guid not in updated_seqr_family_guids:
-                print("Deleting SeqrFamily: %s" % f)
-                f.delete()
+            # delete projects that are in SeqrFamily table, but not in BaseProject table
+            for f in SeqrFamily.objects.all():
+                if f.guid not in updated_seqr_family_guids:
+                    print("Deleting SeqrFamily: %s" % f)
+                    f.delete()
 
-        # delete projects that are in SeqrProject table, but not in BaseProject table
-        for p in SeqrProject.objects.all():
-            if p.guid not in updated_seqr_project_guids:
-                while True:
-                    i = raw_input('Delete SeqrProject %s? [Y/n]' % p.guid)
-                    if i == 'Y':
-                        p.delete()
-                    else:
-                        print("Keeping %s .." % p.guid)
-                    break
+            # delete projects that are in SeqrProject table, but not in BaseProject table
+            for p in SeqrProject.objects.all():
+                if p.guid not in updated_seqr_project_guids:
+                    while True:
+                        i = raw_input('Delete SeqrProject %s? [Y/n]' % p.guid)
+                        if i == 'Y':
+                            p.delete()
+                        else:
+                            print("Keeping %s .." % p.guid)
+                        break
 
         logger.info("Done")
         logger.info("Stats: ")
         for k, v in counters.items():
             logger.info("  %s: %s" % (k, v))
+
+
+def create_sample_records(sample_type, source_project, source_individual, new_project, new_individual, counters):
+
+    vcf_files = [f for f in source_individual.vcf_files.all()]
+    vcf_path = None
+    if len(vcf_files) > 0:
+        # get the most recent VCF file (the one with the highest primary key
+        vcf_files_max_pk = max([f.pk for f in vcf_files])
+        vcf_path = [f.file_path for f in vcf_files if f.pk == vcf_files_max_pk][0]
+
+    if vcf_path:
+        new_sample_batch, sample_batch_created = get_or_create_sample_batch(
+            new_project,
+            sample_type=sample_type,
+            genome_build_id=GENOME_BUILD_GRCh37,
+        )
+
+        sample, sample_created = get_or_create_sample(
+            source_individual,
+            new_sample_batch,
+            new_individual,
+        )
+
+        sample.deprecated_base_project = source_project
+        sample.save()
+
+        if sample_created: counters['samples_created'] += 1
 
 
 def update_model_field(model, field_name, new_value):
@@ -248,8 +293,9 @@ def transfer_project(source_project):
         deprecated_project_id=source_project.project_id.strip(),
     )
     if created:
-        print("Created SeqrSequencingSample", new_project)
+        print("Created SeqrSample", new_project)
 
+    update_model_field(new_project, 'guid', new_project._compute_guid()[:ModelWithGUID.MAX_GUID_SIZE])
     update_model_field(new_project, 'name', (source_project.project_name or source_project.project_id).strip())
     update_model_field(new_project, 'description', source_project.description)
     update_model_field(new_project, 'deprecated_last_accessed_date', source_project.last_accessed_date)
@@ -281,6 +327,7 @@ def transfer_project(source_project):
         except ObjectDoesNotExist as e:
             raise Exception('LocusList "%s" not found. Please run `python manage.py transfer_gene_lists`' % (
                 source_gene_list.name or source_gene_list.slug))
+
         assign_perm(user_or_group=new_project.can_view_group, perm=CAN_VIEW, obj=new_list)
 
     # add collaborators to new_project.can_view_group and/or can_edit_group
@@ -302,17 +349,17 @@ def transfer_family(source_family, new_project):
 
     new_family, created = SeqrFamily.objects.get_or_create(project=new_project, family_id=source_family.family_id)
     if created:
-        print("Created SeqrSequencingSample", new_family)
+        print("Created SeqrSample", new_family)
 
     update_model_field(new_family, 'display_name', source_family.family_name or source_family.family_id)
     update_model_field(new_family, 'description', source_family.short_description)
     update_model_field(new_family, 'pedigree_image', source_family.pedigree_image)
-    update_model_field(new_family, 'analysis_status', source_family.analysis_status)
-    update_model_field(new_family, 'analysis_summary', source_family.analysis_summary_content)
     update_model_field(new_family, 'analysis_notes', source_family.about_family_content)
+    update_model_field(new_family, 'analysis_summary', source_family.analysis_summary_content)
     update_model_field(new_family, 'causal_inheritance_mode', source_family.causal_inheritance_mode)
+    update_model_field(new_family, 'analysis_status', source_family.analysis_status)
     update_model_field(new_family, 'internal_case_review_notes', source_family.internal_case_review_notes)
-    update_model_field(new_family, 'internal_case_review_brief_summary', source_family.internal_case_review_brief_summary)
+    update_model_field(new_family, 'internal_case_review_summary', source_family.internal_case_review_summary)
 
     return new_family, created
 
@@ -322,15 +369,17 @@ def transfer_individual(source_individual, new_family, new_project, connect_to_p
 
     new_individual, created = SeqrIndividual.objects.get_or_create(family=new_family, individual_id=source_individual.indiv_id)
     if created:
-        print("Created SeqrSequencingSample", new_individual)
+        print("Created SeqrSample", new_individual)
 
-    update_model_field(new_individual, 'display_name', source_individual.nickname or source_individual.indiv_id)
     update_model_field(new_individual, 'created_date', source_individual.created_date)
     update_model_field(new_individual, 'maternal_id',  source_individual.maternal_id)
     update_model_field(new_individual, 'paternal_id',  source_individual.paternal_id)
     update_model_field(new_individual, 'sex',  source_individual.gender)
     update_model_field(new_individual, 'affected',  source_individual.affected)
+    update_model_field(new_individual, 'display_name', source_individual.nickname or source_individual.indiv_id)
+    #update_model_field(new_individual, 'notes',  source_individual.notes) <-- notes exist only in the new SeqrIndividual schema. other_notes was never really used
     update_model_field(new_individual, 'case_review_status',  source_individual.case_review_status)
+    update_model_field(new_individual, 'case_review_status_accepted_for',  source_individual.case_review_status_accepted_for)
     update_model_field(new_individual, 'phenotips_eid',  source_individual.phenotips_id)
     update_model_field(new_individual, 'phenotips_data',  source_individual.phenotips_data)
 
@@ -373,6 +422,7 @@ def _retrieve_and_update_individual_phenotips_data(project, individual):
             individual.phenotips_eid,
             is_external_id=True
         )
+
     except phenotips_api.PhenotipsException as e:
         print("Couldn't retrieve latest data from phenotips for %s: %s" % (individual, e))
         return
@@ -381,38 +431,40 @@ def _retrieve_and_update_individual_phenotips_data(project, individual):
 
 
 def get_or_create_sample(source_individual, new_sample_batch, new_individual):
-    """Creates and returns a new SequencingSample based on the provided models."""
+    """Creates and returns a new Sample based on the provided models."""
 
-    new_sample, created = SeqrSequencingSample.objects.get_or_create(
+    loaded_date = look_up_loaded_date(source_individual)
+
+    new_sample, created = SeqrSample.objects.get_or_create(
         sample_batch=new_sample_batch,
         sample_id=(source_individual.vcf_id or source_individual.indiv_id).strip(),
         created_date=new_individual.created_date,
 
         individual_id=source_individual.indiv_id.strip(),
         sample_status=source_individual.coverage_status,
-        bam_path=source_individual.bam_file_path,
-        #picard fields=
+
+        source_file_path=source_individual.bam_file_path,
     )
 
-    new_individual.sequencing_samples.add(new_sample)
+    new_sample.deprecated_base_project=source_individual.family.project
+    new_sample.is_loaded=source_individual.is_loaded()
+
+    new_sample.loaded_date=loaded_date
+    new_sample.save()
+
+    new_individual.samples.add(new_sample)
 
     return new_sample, created
 
 
-def get_or_create_sample_batch(new_project, path, sequencing_type, genome_build_id):
+def get_or_create_sample_batch(new_project, sample_type, genome_build_id):
     new_sample_batch, created = SeqrSampleBatch.objects.get_or_create(
         name=new_project.name,
         description=new_project.description,
         created_date=new_project.created_date,
-        sequencing_type=sequencing_type,
+        sample_type=sample_type,
         genome_build_id=genome_build_id,
     )
-
-    if path is not None:
-        new_sample_batch.variant_callset_path = path
-        new_sample_batch.save()
-
-    # TODO populate is_loaded, load time
 
     if created:
         # SampleBatch permissions - handled same way as for gene lists, except - since SampleBatch
@@ -431,6 +483,8 @@ def get_or_create_variant_tag_type(source_variant_tag_type, new_project):
         name=source_variant_tag_type.tag,
         description=source_variant_tag_type.title,
         color=source_variant_tag_type.color,
+        order=source_variant_tag_type.order,
+        is_built_in=(source_variant_tag_type.order is not None),
     )
 
     return new_variant_tag_type, created
@@ -449,6 +503,7 @@ def get_or_create_variant_tag(source_variant_tag, new_family, new_variant_tag_ty
             alt=source_variant_tag.alt,
             family=new_family,
         )
+        # TODO populate variant_annotation, variant_genotypes
 
         new_variant_tag.search_parameters = source_variant_tag.search_url
         new_variant_tag.save()
@@ -464,11 +519,7 @@ def get_or_create_variant_tag(source_variant_tag, new_family, new_variant_tag_ty
             ref=source_variant_tag.ref,
             alt=source_variant_tag.alt,
             family=new_family,
-            #gene_id=,
-            #transcript_id=, # TODO update gene_id, transcript_id, molecular_consequence
-            #molecular_consequence=
-        )
-
+        ) # TODO populate variant_annotation, variant_genotypes
 
     return new_variant_tag, created
 
@@ -487,9 +538,34 @@ def get_or_create_variant_note(source_variant_note, new_project, new_family):
         alt=source_variant_note.alt,
         search_parameters=source_variant_note.search_url,
         family=new_family,
-        #gene_id=,
-        #transcript_id=, # TODO update gene_id, transcript_id, molecular_consequence
-        #molecular_consequence=,
-    )
+    )  # TODO populate variant_annotation, variant_genotypes
 
     return new_variant_note, created
+
+
+def look_up_loaded_date(source_individual):
+    """Retrieve the data-loaded time for the given individual"""
+
+    # decode data loaded time
+    loaded_date = None
+    try:
+        datastore = get_datastore(source_individual.project.project_id)
+        family_collection = datastore._get_family_collection(
+            source_individual.project.project_id,
+            source_individual.family.family_id
+        )
+        if not family_collection:
+            logger.error("mongodb family collection not found for %s %s" % (
+                source_individual.project.project_id,
+                source_individual.family.family_id))
+            return
+        record = family_collection.find_one()
+        if record:
+            loaded_date = record['_id'].generation_time
+            logger.info("%s data-loaded date: %s" % (source_individual.project.project_id, loaded_date))
+    except Exception as e:
+        logger.error('Unable to look up loaded_date for %s' % str(source_individual))
+        logger.error(e)
+
+    return loaded_date
+
